+++ packages/polaris-viz-core/src/index.ts
@@ -79,6 +79,7 @@ export {
   isGradientType,
   uniqueId,
   createTheme,
+  paddingStringToObject,
 } from './utilities';
 export {
   useSparkBar,