+++ packages/polaris-viz-native/src/components/ChartContainer/ChartContainer.tsx
@@ -1,14 +1,18 @@
 import React, {useState, ReactElement, cloneElement} from 'react';
 import {View} from 'react-native';
-import type {Dimensions} from '@shopify/polaris-viz-core';
-import {useTheme} from '@shopify/polaris-viz-core';
+import {
+  Dimensions,
+  paddingStringToObject,
+  useTheme,
+} from '@shopify/polaris-viz-core';
 
 interface Props {
   children: ReactElement;
   theme?: string;
+  sparkChart?: boolean;
 }
 
-export function ChartContainer({theme, children}: Props) {
+export function ChartContainer({theme, children, sparkChart = false}: Props) {
   const [chartDimensions, setChartDimensions] =
     useState<Dimensions | null>(null);
 
@@ -16,21 +20,34 @@ export function ChartContainer({theme, children}: Props) {
 
   const handleOnLayout = (event) => {
     const {width, height} = event.nativeEvent.layout;
-    setChartDimensions({width, height});
+
+    setChartDimensions((previousDimensions) => {
+      if (previousDimensions?.width === width) {
+        return {
+          width: previousDimensions?.width,
+          height: previousDimensions?.height,
+        };
+      }
+      return {width, height};
+    });
   };
+
   return (
     <View
       onLayout={handleOnLayout}
       style={[
         {
           borderRadius: Number(parseInt(chartContainer.borderRadius, 10)),
-          padding: Number(parseInt(chartContainer.padding, 10)),
           backgroundColor: chartContainer.backgroundColor,
+          ...paddingStringToObject(chartContainer.padding),
         },
         {
           position: 'relative',
           width: '100%',
           height: '100%',
+          minHeight: sparkChart
+            ? chartContainer.sparkChartMinHeight
+            : chartContainer.minHeight,
         },
       ]}
     >
@@ -38,7 +55,10 @@ export function ChartContainer({theme, children}: Props) {
         ? null
         : cloneElement(children, {
             theme,
-            dimensions: chartDimensions,
+            dimensions: {
+              width: chartDimensions.width,
+              height: chartDimensions.height,
+            },
           })}
     </View>
   );