+++ packages/polaris-viz/src/components/BarChart/stories/BarChart.stories.tsx
@@ -14,6 +14,7 @@ import {
 } from '../../../storybook';
 
 import {generateMultipleSeries} from '../../Docs/utilities';
+import {PageWithSizingInfo} from '../../Docs/stories/components/PageWithSizingInfo';
 
 const TOOLTIP_CONTENT = {
   empty: undefined,
@@ -139,9 +140,10 @@ export default {
   parameters: {
     horizontalMargin: 0,
     docs: {
+      page: PageWithSizingInfo,
       description: {
         component:
-          'Used to show comparison of different types, across categories or time. Bars can be stacked or side by side. It is recommended that you use a legend whenever displaying multi-series data. To display one, use the `<Legend />` component. <br /> <br /> This component inherits its height and width from its container.',
+          'Used to show comparison of different types, across categories or time. Bars can be stacked or side by side.',
       },
     },
     controls: {